feat: add custom dev proxy vite plugin & move `fetchApi` into `Api` model
