+++ package.json
@@ -32,8 +32,8 @@
   "devDependencies": {
     "@commitlint/cli": "19.3.0",
     "@commitlint/config-conventional": "19.2.2",
+    "@majority-elite/remix-dev-utils": "0.1.2",
     "@remix-run/dev": "2.11.2",
-    "@remix-run/node": "2.11.2",
     "@remix-run/testing": "2.11.2",
     "@storybook/addon-essentials": "8.2.9",
     "@storybook/addon-interactions": "8.2.9",
@@ -62,7 +62,6 @@
     "husky": "9.1.1",
     "lint-staged": "15.2.7",
     "prettier": "3.3.3",
-    "set-cookie-parser": "2.7.0",
     "storybook": "8.2.9",
     "typescript": "5.5.3",
     "vite": "5.3.3",
