+++ server/utils/cloudflare.ts
@@ -5,7 +5,6 @@ import {
 import { type GetLoadContextFunction } from '@remix-run/cloudflare-pages';
 import { type PlatformProxy } from 'wrangler';
 import { clientEnvSchema, serverEnvSchema } from '../constants/env';
-import type { FetchApi } from '../types/api';
 import type { AuthSession, AuthSessionData } from '../types/auth';
 import type {
   ClientEnv,
@@ -13,49 +12,18 @@ import type {
   ContextEnv,
   ServerEnv,
 } from '../types/env';
-import { fetchApiImpl } from '../utils/api';
 
 declare module '@remix-run/cloudflare' {
   interface AppLoadContext extends ServerEnv, CloudflareEnv {
     readonly clientEnv: ClientEnv;
     readonly authSession: AuthSession;
-    /**
-     * `ApiInfo` 객체로 백엔드 API를 호출하는 함수
-     * @example
-     * ```
-     * import { type ActionFunctionArgs } from '@remix-run/cloudflare';
-     * import { api_getPost } from '@/apis/post';
-     *
-     * export const action = async ({ context }: ActionFunctionArgs) => {
-     *   const apiReturn = await context.fetchApi(api_getPost, { id: 4 });
-     *   if (!apiReturn.isSuccess) {
-     *     return apiReturn.errorResponse;
-     *   }
-     *   const { response } = apiReturn;
-     *   // ...
-     * };
-     * ```
-     * - `loader`와 같이 오류 발생 시 `ErrorBoundary`로 이동해야 하는 경우 `throwOnError` 옵션 사용
-     * ```
-     * import { type LoaderFunctionArgs } from '@remix-run/cloudflare';
-     * import { api_getPost } from '@/apis/post';
-     *
-     * export const loader = async ({ context }: LoaderFunctionArgs) => {
-     *   const { response } = await context.fetchApi(
-     *     api_getPost,
-     *     { id: 4 },
-     *     { throwOnError: true },
-     *   );
-     *   // ...
-     * };
-     * ```
-     */
-    readonly fetchApi: FetchApi;
+    readonly commitSession: () => Promise<string>;
   }
 }
 
 export const getLoadContext: <Cf extends CfProperties>(
   authSession: AuthSession,
+  commitSession: () => Promise<string>,
   args: {
     request: Request;
     context: {
@@ -64,6 +32,7 @@ export const getLoadContext: <Cf extends CfProperties>(
   },
 ) => ReturnType<GetLoadContextFunction<ContextEnv>> = async (
   authSession,
+  commitSession,
   { context },
 ) => {
   const clientEnv = clientEnvSchema.parse(context.cloudflare.env);
@@ -73,24 +42,14 @@ export const getLoadContext: <Cf extends CfProperties>(
     ...context.cloudflare.env,
     clientEnv,
     authSession,
-    fetchApi: (async (api, variables, options) =>
-      fetchApiImpl(
-        api,
-        variables,
-        context.cloudflare.env.API_URL,
-        authSession,
-        options,
-      )) as FetchApi,
+    commitSession,
   };
 };
 
-export const makeAuthSession = async (
-  env: {
-    authCookieSessionSecret: string;
-    kvNamespace: KVNamespace<string>;
-  },
-  cookieHeader?: string | null,
-) => {
+export const makeAuthSessionStorage = async (env: {
+  authCookieSessionSecret: string;
+  kvNamespace: KVNamespace<string>;
+}) => {
   const authSessionCookie = createCookie('__auth_session', {
     secrets: [env.authCookieSessionSecret],
     sameSite: true,
@@ -99,7 +58,6 @@ export const makeAuthSession = async (
     cookie: authSessionCookie,
     kv: env.kvNamespace,
   });
-  const authSession = await authSessionStorage.getSession(cookieHeader);
 
-  return { authSessionStorage, authSession };
+  return authSessionStorage;
 };
