--- server/vite/index.ts
@@ -1,1 +1,0 @@
-export { default as remixCloudflareDevProxyVitePlugin } from './remixCloudflareDevProxyVitePlugin';
